Tested BIDS version 1.2.0

+++ run_xasl.py
@@ -29,15 +29,15 @@ print('Parsing module inclusion parameters...')
 import_modules = ""
 process_modules = ""
 
-if args.mode == "BIDS_Subject":
+if args.mode == "BIDS":
 
     import_modules = f"[0,0,{args.im_deface},1]"
     process_modules = f"[{args.pm_structural},{args.pm_asl},0]"
 
-if args.mode == "BIDS_Dataset":
+# if args.mode == "BIDS_Dataset":
 
-    import_modules = f"[0,0,{args.im_deface},1]"
-    process_modules = f"1"
+#     import_modules = f"[0,0,{args.im_deface},1]"
+#     process_modules = f"1"
 
 if args.mode == "DICOM":
 
@@ -66,7 +66,18 @@ if args.mode == "BIDS":
 
     print('Copying input files...')
     print('Copying contents of ' + args.input_folder + ' to ' + dest_path)
-    shutil.copytree(args.input_folder, dest_path)
+
+    # For development purposes:
+    # When running outside CBRAIN (e.g. directly in Docker), input will have one less layer
+    # Check for dataset_description.json; if not present, add extra folder layer and copy dataset_description_json
+
+    if os.path.isfile(os.path.join(args.input_folder, "dataset_description.json")):
+        print("BIDS Dataset found.")
+        shutil.copytree(args.input_folder, dest_path)
+    else:
+        print("BIDS Subject found. Copying dataset_description.json")
+        shutil.copytree(args.input_folder, os.path.join(dest_path, args.input_folder))
+        shutil.copy(args.dataset_description_json, dest_path)
 
 ## DICOM:
 ## Should already have correct folder structure + all contained JSONs!
